update GH unit test workflow DB (#27)
## Description
The latest version of pyway, 0.3.30, is not able to read the existing
migrations directory. Since these migrations are deprecated in favor of
alembic, the easiest path forward is to pin pyway to version 0.3.29
until its removed from the project.

Within #19 we simplified the settings to use the new DB_URI env variable
to define connection settings. We forgot to also update postgres
database setup in the GH testing workflow to follow this pattern.