seperating matching functions into its own module (#20)
## Description
Moving matching functions into its own module so they can be more easily
reused. Keeping these functions separate, from the logic used for
linkage, allows us to move forward with creating an alternative set of
linkage functions using a new schema. Also, changing the syntax of the
algorithm configuration for all value references that map to python
functions. The new format is `func:<module_path>.<function_name>`, this
not only makes it easy for parsing the algorithm, but also makes it
transparent to the author as to where the matching code lives.

## Related Issues
closes #9