remove old schema and code (#68)
## Description
5000 lines removed 😮 

## Related Issues
#63 

## Additional Notes

- Making SQLite the default database for testing and local_server, now
the old code is removed. HOWEVER, we will continue to use postgres in CI
- Removed assets and migrations related to old schema
- explicitly added DB adapters (postgres, mysql, microsoft sql server)
to pyproject.toml
- consolidating testing session fixtures by creating a new
contextmanager function, `get_test_session`, in database.py and adding
[conftest.py](https://docs.pytest.org/en/7.1.x/reference/fixtures.html#conftest-py-sharing-fixtures-across-multiple-files)
to assist with creating reusable testing fixtures
- removed all `recordlinker.linkage` code and tests, except for
matchers, this was copied to `recordlinker.linking.old_matchers` (we
need to port these methods before we can move them)
- removed old schema calls from the link-record API endpoint
- removed pyway helpers from `recordlinker.utils`
- renamed test_record_linkage to test_main (as its really just testing
the API endpoints in main.py). Rewrote some of the tests to make use of
the new client fixture.