Optimizing initialization of evaluators and rule
Binding the AlgorithmPass evaluators and rule to functions for every
compare call adds overhead.  Its quicker if we just do that once for the
pass, and allow all subsequent calls to reuse the bound values.

This PR adds new bound_evaluators and bound_rule functions to the
AlgorithmPass model so we can cache the binding operations for future
use.