Optimizing initialization of evaluators and rule (#77)
## Description
Adds the `bound_evaluators` and `bound_rule` functions to the
AlgorithmPass model so we can cache the binding operations for future
use.

## Related Issues
This is a small optimization for features added in #72 

## Additional Notes
Binding the AlgorithmPass evaluators and rule to functions for every
compare call adds overhead. It's quicker if we just do that once for the
pass, and allow all subsequent calls to reuse the bound values.