project structure cleanup (#69)
## Description
Restructuring the codebase into `models/` and `schemas/`. The former
will be used for [sqlalchemy ORM models](https://www.sqlalchemy.org/)
and the latter for [pydantic validation
models](https://docs.pydantic.dev/latest/). This is necessary so we
create a strategy for NOT having circular imports.

## Related Issues
refs #65 

## Additional Notes
With the separation of models into 2 distinct packages,
`recordlinker.models` and `recordlinker.schemas`, we will allow schemas
to depend on models, but not the other way around.

- `BlockingKeys.to_value` was migrated to `PIIRecord.blocking_keys` to
resolve the dependency issue
- There is still an issue with `Patient.record`, as this getter/setter
has a dependency on the. I'm not sure the best way to address this. The
code really isn't necessary, but its an optimization right now that is
helpful. I'm going to add an [issue to resolve this in the
future](https://github.com/orgs/CDCgov/projects/34/views/1?sortedBy%5Bdirection%5D=asc&sortedBy%5BcolumnId%5D=Assignees&pane=issue&itemId=82498554).
Right now we can circumvent and potential problems by moving the import
inside the function, but that is not ideal.