Auto initialize algorithms in DB (#89)
## Description
Configuring the code to optionally load in initial algorithms after the
tables have been initialized. Doing so, removes our needs now for having
to use a database migration tool for the near future.

## Related Issues
closes #57 

## Additional Notes
- This is contentious, but have removed alembic. Using alembic has
required a lot of manual effort to get migrations correct, with little
benefit this early in a project. Our schema hasn't stabilized yet, and
we don't have any systems running the code in production, so it's of
little value. Over time, we may need to bring this back to safely adjust
production schemas, but now is not the time.
- Created `assets/initial_algorithms.json` with the two existing DIBBs
algorithms. We can use this for our test cases, and also initializing a
new database with the data, if the INITIAL_ALGORITHMS env variable is
set. This also gives customers another option for managing their
algorithms, in they can write their own and set the INITIAL_ALGORITHMS
env variable to load that json file.
- Capitalized `schemas.pii.Features`. The blocking keys were
capitalized, so making them both capitalized will hopefully reduce
errors in writing the algorithm specs.
- Consolidated some of the assets and testing assets, as there were
duplicates and there was some testing files in the top level one.

Co-authored-by: Alex Hayward <44384574+alhayward@users.noreply.github.com>