endpoints to update person cluster on patient (#101)
## Description
Adding two new endpoints to support updating the Person cluster a
Patient belongs to. One is for assigning the Patient to a specific
cluster, and another is for assigning them to a new cluster.

## Related Issues
closes #96 

## Additional Notes
- `POST /patient/<ref-id>/person` takes an empty payload, creates a new
Person cluster, assigns the Patient to that cluster and returns back
reference ids for both the Patient and Person.
- `PATCH /patient/<ref-id>/person` takes a person ref id payload,
assigned the Patient to that existing Person and returns back reference
ids for both the Patient and Person.