splunk hec logs (#107)
## Description
Adding a new SplunkHandler that can emit logs to a Splunk HEC instance
if configured.

## Related Issues
closes #83 

## Additional Notes
The SplunkHandler is using threading to make the actual send to the HEC.
This is to not block on the IO of sending the data over the network.
Instead we just fire the log entry over in the background and move
forward. Now there is a risk that the data never makes it to the HEC,
and we don't know about it, but for this use case that is an acceptable
risk.