seed API (#124)
## Description
Adding a new API endpoint, `/seed`, to control initializing the MPI
tables with data. `POST /seed` is used to insert person/patient clusters
into the MPI and `DELETE /seed` is used to reset the MPI.

## Related Issues
closes #122

## Additional Notes
- Updating the unit test CI job to combine all the coverage reports from
the matrix of tests before uploading to codecov. This changeset adds
some features not compatible with MySQL, thus some test cases are
skipped depending on the database dialect.
- adding `mpi_service.bulk_insert_patients`, essentially the same
functionality of insert_patient, but is faster when attempting to insert
more than 1 patient.
- [MySQL doesn't support all the bulk insert
features](https://docs.sqlalchemy.org/en/20/orm/queryguide/dml.html#getting-new-objects-with-returning),
so it will be slower than the other dialects when using /seed