fix misleading matcher names (#134)
## Description
Renaming `Feature.TELEPHONE` to `Feature.TELECOM`, as that more
accurately describes the matching behavior. Adding` Feature.GIVEN_NAME`
and updating `Feature.FIRST_NAME` to only use the first element of a
given name.

## Related Issues
closes #128
closes #129 

## Additional Notes
- TELEPHONE was misleading because the telecom field could collect
emails, fax numbers and other contact points, that would be included in
the matches. Renaming this to TELECOM better explains to the user what
we're matching on.
- FIRST_NAME was misleading, as it was including the middle names in its
matches as well.