bugfix: bulk insert serialization of data (#136)
## Description
Data can be inserted when linking or seeding, in the former we were
inserting the data as a dictionary into the Patient.data column, however
in the latter we inserted the data as a string. This causes issues on
subsequent reads when using the data.

The call to convert the PIIRecord into a value for the database has been
standardized in PIIRecord.to_dict(), and the bulk insert method (which
is used in the seeding process) has been updated to use this method.