Ability to delete a Patient record (#141)
## Description
Mistakes can happen, and there could be times when customers send a
record to the link API by mistake (could be a known duplicate document,
or just an error in the workflow). We should have the ability to "undo"
those mistakes. Create a new DELETE /patient/<patient_reference_id>
endpoint that will allow users to remove a known patient record.

## Related Issues
closes #138 

## Additional Notes