NBS Test Cases  (#123)
## Description
NBS provided us a spreadsheet that defines an existing state and then a
bunch of tests with expected results. Architect a testing script of
sorts, that can parse a similar spreadsheet (can just be a CSV) with
seeding data and iterate through the tests with the expectations.

## Related Issues
closes #110 

## Additional Notes
For a full explanation of run instructions, directory structure, and
file uses check out the
[README.md](https://github.com/CDCgov/RecordLinker/pull/123/files#:~:text=%23%20Record%20Linkage%20Algorithm%20Testing)

The key inputs to these tests will be the following:

- a CSV file that can be used to seed the MPI
- a CSV with test cases (basically the same format as the above), but
also a column to indicate if it is a match and to whom
- a JSON file with an algorithm config to use
The setup should make it easy to edit any 3 of those above files, run
the test cases, and examine the results.