fix: docker builds with libodbc (#179)
## Description
Corrects the syntax of the if block in the Dockerfile to ensure proper
evaluation for installing msodbcsql18 and unixodbc-dev. This fix
resolves an issue where the Record Linker image fails to start when
configured with a sqlserver DB_URI.

## Additional Notes
This is a significant issue for users who want to: a) use our published
Docker images on ghcr.io and b) connect to a sqlserver database. The if
block on line 30 of the Dockerfile has syntax that works with /bin/bash
but not with /bin/sh. As a result, the step was always skipped during
the image build process.

This update introduces a database matrix to our smoke tests, enabling
the Docker image to be tested against all four supported databases, not
just PostgreSQL.