new local scheme of dirty-tag for developer installation (#191)
## Description
When we install RecordLinker into an environment, a version number is
automatically calculated from git data using setuptools scm. If the HEAD
is a version tag, then the version is automatically calculated to be
that. In development environments, though, we take the last version and
add a "post.0" to it, showing it's a new release post the last named
one.

Additionally, we were previously calculating a timestamp to add to that
"post.0" value based on when HEAD was committed. This worked most of the
time, however if the developer has a dirty index, the wheel wouldn't
build correctly. We don't need that, and it's fine if a developer tries
to install / reinstall RecordLinker in an environment with a dirty
index. So to get around that, we'll stop adding the timestamp to the end
of "post.0" and instead just call it "post.0+dirty" (if the index is
indeed dirty).