correct 508 issues on site docs (#194)
## Description
Making small CSS changes to account for 508 issues
[axe](https://www.deque.com/axe/) flagged on the site docs. Note: this
does not account for any issues on the API docs.

## Notes
The syntax highlighting in the code blocks can cause readability issues
due to poor contrast. One way to correct for this is to remove the
syntax highlighting (just mark every code block as plaintext).