Reverse identifiers for comparisons (#211)
## Description

Since Jaro-Winkler is the default similarity measure for string
comparisons, more weight is given to differences at the start of a
string. As such, this PR modifies the way that we yield and store
Identifiers for comparison. Previously, the Identifier format was
`type:authority:value` but now it is `value:authority:type` to give more
weight to the value. For example, when comparing two MRNs with values,
`1234` and `2345`, we will now give more weight to `1234` vs `2345` than
the type, e.g., `MR` vs `MR`. This should yield better differentiation
between strings.


## Related Issues
Fixes #200 

## Additional Notes
I updated the language in the RFC to reflect the new format and added a
note at the bottom to indicate that the RFC has been changed.