Add endpoint for deleting orphaned persons (#221)
## Description
This PR adds an API endpoint for deleting persons who have no associated
patients. It works by checking that the provided person_reference_id is
valid and then checking if the person has any associated patients.
Persons with associated patients are not deleted.

## Related Issues
Addresses 1 of the 3 API endpoints listed in #163

## Additional Notes
The other API endpoints in #163 will be in follow-up PRs.