Handle missing values in blocked MPI records (#242)
## Description
This PR adds a filter function to the `get_block_data` function in the
`mpi_service`. This filter function removes from consideration all
candidates from the MPI who are part of a person cluster containing a
patient who satisfied blocking criteria, but which themselves have
present values (i.e. are not missing fields) that disagree with incoming
blocking keys. This will make matching more precise and reinforce
validity in the blocking step.

## Related Issues
#231 

## Additional Notes
N/A