Add caching of RecordLinker image to smoke tests (#261)
## Description
This PR addresses half of the caching issue outlined in ticket 224:
ensuring that we use the cached the RecordLinker image when we run smoke
tests to avoid hitting the limit on requests to DockerHub.

After much tinkering, I added a `build-push-run` step to our smoke tests
so now when you run them you should see that we use the image from ghcr
for each smoke test instead of hitting
`docker.io/library/python:3.12-slim` to build the service in each run.

## Related Issues
Fixes #224 

## Additional Notes
Getting the services to run from images in the GHCR has proven to be
much more difficult, which is why I opted to split up the work on #244.
At least we will have cut down the number of requests for RecordLinker
by the 4/1 deadline.

<--------------------- REMOVE THE LINES BELOW BEFORE MERGING
--------------------->

## Checklist
Please review and complete the following checklist before submitting
your pull request:

- [ ] I have ensured that the pull request is of a manageable size,
allowing it to be reviewed within a single session.
- [ ] I have reviewed my changes to ensure they are clear, concise, and
well-documented.
- [ ] I have updated the documentation, if applicable.
- [ ] I have added or updated test cases to cover my changes, if
applicable.
- [ ] I have minimized the number of reviewers to include only those
essential for the review.

## Checklist for Reviewers
Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.