Handle missing fields during evaluation (#259)
## Description
This PR adds more robust handling of comparisons between records in
which one or more fields are missing data. If only a small amount of
data is missing (fewer log-odds-worth than a specified percentage of the
total possible), then missing comparisons are awarded partial weight to
allow for data recovery and quality matches. However, if too much data
is missing, the candidacy is automatically rejected.

## Related Issues
Closes #235 

## Additional Notes
Just changed some variable names from the ticket for clarity