missing blocking keys (#254)
## Description
Changed the get blocking data call to allow for some missing blocking
values. If too many values are missing from the query (indicated by
checking log odds values), then the blocking pass will be skipped.

## Related Issues
closes #230

## Additional Notes
A couple of things to note on the implementation.

1. The naming conventions for the new "compare_minimum_percentage" value
does not match the issue. This was done because the issue was written
with the assumption that #223 would be completed first, which it is not.
The plan is to make changes to #223 to adjust the location of this new
parameter, but for now we can just continue to use `kwargs`.
2. `get_block_data` was converted into a class, `GetBlockData`. The
[Improve Evaluation of Missing
Values](https://github.com/CDCgov/RecordLinker/issues#workspaces/dibbs-record-linker-678acc70237f5000299a9d71/epics/Z2lkOi8vcmFwdG9yL1plbmh1YkVwaWMvMjI2OTc0)
epic has introduced extra conditions for retrieving blocking data.
Checking for those extra conditions requires reusing variables, or at
least benefits from reuse, thus storing some sort of state between the
different functions has an advantage. There are some alternatives, that
are worth considering, that I experimented with before landing on this
solution. I think all three implementations can do a good job of
encapsulating the logic and optimizing the looping constructs for
efficient evaluation. If anyone thinks another implementation would read
better, please voice that.
    - A long function with all the logic encapsulated
    - A long function with nested functions