address line suffix normalize (#258)
## Description
Add a normalization step for normalizing street suffixes and updating
the state normalization process to happen during ingestion.

## Related Issues
closes #239 

## Additional Notes
Previously, state values were being normalized in the feature_iter. In
alignment with the direction we're taking on the other fields, it makes
sense to do this normalization just once, at ingestion time. One key
difference with this, is we are now preserving any invalid state values,
mainly to preserve data when a state could be misspelled.