Add AlgorithmPass label and description fields (#272)
## Description
Add label and description fields to Algorithm Pass to assist users in
identifying which pass they are viewing/editing.

## Related Issues
closes #246 

## Additional Notes
Both of these fields are optional, in the event that a label isn't
specified, we'll default to a value that is based on the evaluators used.
This wasn't specified in the spec, but I thought it be nice to auto
generate some labels for users. I believe Henry was doing something
similar in the mockups he put together for NBS.