dynamic next urls for patient/person routers (#276)
## Description
removing the hard-coded next url paths for our orphaned patient and
person routes, and replacing with versions that are dynamically
generated. The work on the roadshow concept is requiring us to
reorganize our API url paths, which will in effect change the path for
both of these routes. Doing this dynamically makes it easier to
encapsulate all those url path changes in main.py.