named routes (#277)
## Description
Name all routes to create flexibility in testing and future
configuration.

## Related Issues
refs #266 

## Additional Notes
Naming our routes gives us the ability to remove hardcoded paths in our
routing test cases, which opens the door for future flexibility in our
API route paths. A couple of places where this could be useful

- In our roadshow demo, we are planning on serving the UI from the root
path (eg `/`, `/wizard`, etc). This would conflict with our API routes,
so adjustments may need to be made in some environments.
- NBS has requested the desire to change their routes to align with
routing standards in their K8 cluster (eg `/` -> `/recordlinker/api`).
If we want to give users the ability to adjust this through
configuration, it'll be best to keep our routing paths in our test cases
dynamic so it will work with these new settings.