Use schemas.Algorithm for linking (#321)
## Description
Send a value of `schemas.Algorithm` and not `models.Algorithm` when
linking records.

## Related Issues
closes #294 

## Additional Notes
When we save the Algorithm configuration to the database, we store many
values as strings (eg blocking keys, features, functions, etc). However,
when using the configuration during linkage, we prefer using the enum,
functions and objects those strings represent. Previously, this meant we
have to cast the string values to its native type in the link and
compare methods.

By converting the model data back to a schema.Algorithm object, we can
do that casting one time before the linkage process kicks off. This
helps simplify our linkage logic a bit as it no longer needs to worry
about how to cast strings back to their native data types.