Add suffix normalization (#322)
## Description
This changeset adds a pydantic validator that parses and normalizes
suffixes according to a predefined mapping. I've left the mapping in an
assets file in case we want to add roman numeral standardization at some
later date.

## Related Issues
Fixes #297 

## Additional Notes
I did some research on common suffixes in names in the US, and most
results indicated it's pretty limited, just Junior, Senior, II, and III.
It's easy to support roman numerals, as well as any other spelling or
punctuated variants thereof, but given this I think leaving the suffixes
we intend to support in an assets file (like with state abbreviations
and street conventions) follows our established practice.