Prepend during blocking (#327)
## Description
This PR prepends normalized suffixes to first names during blocking.
Following the convention of the existing `feature_iter` entry for First
Name, only the first suffix is pulled, and _only_ if it has been
normalized as a result of parsing. The least invasive way I found to
make this change was an optional boolean flag on feature_iter proper,
since otherwise in the blocking keys call we can't call feature_iter on
First Name and have to basically repeat / reconstruct that code, but if
that's preferable that's also okay with me.

## Related Issues
Closes #298