Match review page  (#329)
## Description
This work implements the match review page with mock content.
 
- Adds the `match-review` route in FastAPI.
- Creates `MatchReview` page with static elements. The id parameter
needs to be passed as a query parameter because exporting the nextjs app
as static assets does not support dynamic routes (e.g.:
/match-review/<id>).
- Creates `RecordReview` component that handles the retrieval of data
and rendering of the record table and the `RecordCompare`.
- Creates `RecordCompare` component which is the section that list all
the fields and values that user needs to review.
- Reworks the ServerError component to make it work for the match review
page width.
- Introduces reusable icons implementation.
- Updates the current models to match better what the endpoints will
send, still needs to be polished more once backend integration is worked
on.
- Updates naming across the app elements to match better the terminology
in figma.
- Disables rule react-hooks/exhaustive-deps so we can set useEffects to
trigger only in initialization.

## Related Issues
Implements #306


![localhost_3000_match-review_id=123](https://github.com/user-attachments/assets/261436df-6e4e-4825-8e7f-f2c619f9abe6)

---------

Co-authored-by: Eric Buckley <eric.buckley@gmail.com>