schema.AlgorithmContext (#330)
## Description
Create AlgorithmContext nested schema object to store
include_multiple_results, log_odds and skip_values.

## Related Issues
closes #293 

## Additional Notes
The `max_log_odds_points` parameter was removed from the compare method.
It's still required in the function, but now we need to retrieve the
log_odds value for a feature BEFORE we call the matcher function
(previously the matcher functions were retrieving this from the kwargs).
Since we're doing that in compare, it seemed simple enough to just make
that a calculated value (summations are simple and quick).