Add link and unlink endpoints (#335)
## Description
This PR adds 2 new endpoints: 
1) "link" endpoint to update the `linked_status` to `True` when a user
clicks on the "link" button. It also updates the `person_id` of the
incoming record to match the `person_id` of the potential match record.
The data is also stored in the session.
2) "unlink" endpoint to update the `linked_status` to `False` when a
user clicks on the "unlink" button. It also updates the `person_id` of
the incoming record to `null`. The data is also stored in the session.

## Related Issues
Closes #313 

## Additional Notes
I also removed the `max_age` of a cookie since we will be handling the
reset in a follow up ticket.