Add stats calculations and results display (#342)
## Description
This PR is a fast-follow to a conversation had on 4/30 in which we
established we wanted a standardized performance calculation script to
be uploaded to our repo. This script integrates into the `run_test.py`
script executed as part of algorithm tests, and gives us calculations
for: raw counting stats (TP, FP, TN, FN), derived match quality stats
(sens, spec, F1, PPV), and labels for matches graded possible by either
system. This should standardize the way we calculate and report
performance going forward.

**NOTE:** The change numbers look massive, but that's because we got the
all-clear to upload our more robust test and seed data sets into the
algorithm test configuration. There are only a couple changed code
files, most of this is CSV uploading.