Handle suffix during evaluation (#346)
## Description
This PR adds the final piece of handling suffixes during name
evaluation. Previously, we used them only during blocking, but this
change set now includes handling during evaluation as well. In order to
still maintain the First_Name field as its own readable entity, we've
simply re-used the `should-prepend` parameter in `feature_iter`. That
way, it's only during the very specific cases we've outlined that we
combine a name and a suffix.

## Related Issues
Closes #299