Added gzip compression middleware (#352)
## Description
Static bundles should be served as gzip compressed assets whenever the
browser accepts the format.

## Related Issues
None

## Additional Notes
I briefly looked into how gzip compression can be supported and the only
thing I was able to find was on the fly. I anyone knows a better way to
setup this in FastAPI is greatly appreciate it.