Fix bug in scripts/gen_seed_data.py (#360)
## Description
Somewhere in our updates to `PIIRecord`, we change the expectation for
Race to a string. Faker produces a Race `object`, which I changed to a
string so that we can continue to run `gen_seed_data.py`.

Original error:
```
pii.py", line 411, in parse_race
    return [Race.parse(v) for v in value]
                                   ^^^^^
TypeError: 'Race' object is not iterable
```

## Related Issues
Fixes #359 

## Additional Notes
I'm guessing this script will be useful for jurisdictions testing out
Record Linker so I wanted to make sure it stayed current.