remove context from results in algorithm list
The algorithm context was never meant to be in the results of the
listing API call, this call is supposed to just give high level details
of what algorithms are available.  The specific detail API endpoints for
each algorithm, should be used to get info on the algorithm context
and/or the passes.