Add initial locustfile for load testing (#380)
## Description
This PR adds an initial `locustfile` for running load testing initially.
It will need to be modified as the load testing plan develops but this
gives you an idea of how load testing might work locally.

**How to run the load test**:
For now (eventually will make this a single script):
1. Spin up the API locally. I used `./scripts/local_server.sh` 
2. Run `locust -f ./tests/load/locustfile.py --headless -u1 -r 1
--run-time 3s --host http://localhost:8000`

**Interpreting the output**:
You will get two sets out output after the load test has completed:
1. # of requests, # fails, and time for each endpoint that is hit
2. Approximate response time percentiles 

We can save the output as a CSV for easier sharing with `--csv example`

**Details about locustfile**:
1. The seed data is loaded: I went with loading a simple file for now,
but as we increase the number of records, we may need to read in the
file incrementally in case it is too large to open in memory or generate
the seed data on the fly.
4. On start, the seed data is added into the MPI: currently this runs
once per user (not once per test) and will need to be fixed in future
iterations. To avoid seeding multiple times, I have set the number of
users to 1.
5. Randomize the seed data: A random record from the seed data is chosen
and run through record linkage by hitting the `/link` endpoint. In
future iterations, the record will be duplicated a random number of
times and randomized once I have updated `expand_test_data` to work with
`PIIRecords` (ticket #376)


## Related Issues
Closes #377 

## Additional Notes
There are likely lots of iterations we can make on this file as we learn
more about how locust works so let me know if you find something neat!