Accept proxy headers (#386)
## Summary
Add a flag to uvicorn to allow proxy headers, specifically
X-Forwarded-For and X-Forwarded-Proto, from the upstream server. This
will allow for SSL termination to happen upstream and any info necessary
for Record Linker to be preserved in request headers. Which is necessary
when setting a secure session cookie.

## Additional Notes
[Proxies and forwarded
headers](https://www.uvicorn.org/deployment/#proxies-and-forwarded-headers)