Upgrade testing framework (#385)
## Description
This PR provides an overhaul and update to the Record Linker algorithm
match performance testing suite. It provides a robust new source of
testing data--combining regraded NBS cases with new cases designed to
test feature missingness and data quality limitations, as well as
corrects a deficiency in our existing seed file where the last digit of
SSN was getting chopped off. The new included README summarizes the
purpose of different testing files, but for more details on the actual
development plan and interpretation of new testing results, see the
associated document.

## Related Issues
Closes #368