Add scrambler for load testing (#384)
## Description
This PR adds a script to scramble generated records for the purposes of
load testing. It re-uses much of the code @bamader wrote for the
scrambling the csv of NBS data, and adapts it for the json/PIIRecord
data we have. To test out the code, you can run `python3 -m
tests.load.scripts.scramble_data
--file="./tests/load/assets/test_data.json"` and it will spit out a file
with scrambled records and the original from the `test_data.json` file.

## Related Issues
Closes #376 

## Additional Notes
If there is time after load testing to refactor this, there's a good
amount of cleanup/refactoring I'd like to do to make this more adaptable
to other file types, e.g., the NBS CSV and other CSVs. But, given the
priority for load testing (not scrambling), this script does the job and
does it quickly.

Background for why we need the script:
We want to be able to seed the test MPI during load testing with a
variety of records that may or may not match each other. We plan to:
1. generate 1.5 million records (the "original" data)
2. grab ~1 million of the records from the original data and
scramble/multiple them using this script to get roughly 4-5 million
records (the "scrambled" data)
3. Seed the MPI with the scrambled data
4. Randomly select records from the original data to hit the `/link`
endpoint with to see how the database performs with a larger MPI